# chore: Add perf benchmarks