--- a/sandbox/Benchmark/Categories.cs
+++ b/sandbox/Benchmark/Categories.cs
@@ -1,0 +1,7 @@
+﻿namespace Benchmark;
+
+public static class Categories
+{
+    public const string LINQ = "LINQ";
+    public const string ZLinq = "ZLinq";
+}