--- a/tests/ZLinq.Tests/Linq/MaxByTest.cs
+++ b/tests/ZLinq.Tests/Linq/MaxByTest.cs
@@ -1,4 +1,4 @@
-#pragma warning disable
+﻿#pragma warning disable
 namespace ZLinq.Tests.Linq;
 
 public class MaxByTest
@@ -33,10 +33,10 @@ public class MaxByTest
     public void IntegerSourceWithIntegerKey()
     {
         var ints = new[] { 5, 3, 8, 2, 10 };
-        
+
         ints.AsValueEnumerable().MaxBy(x => x).ShouldBe(ints.MaxBy(x => x));
         ints.ToValueEnumerable().MaxBy(x => x).ShouldBe(ints.MaxBy(x => x));
-        
+
         // Using a different key selector
         ints.AsValueEnumerable().MaxBy(x => -x).ShouldBe(ints.MaxBy(x => -x));
         ints.ToValueEnumerable().MaxBy(x => -x).ShouldBe(ints.MaxBy(x => -x));
@@ -46,7 +46,7 @@ public class MaxByTest
     public void IntegerSourceWithStringKey()
     {
         var ints = new[] { 5, 3, 8, 2, 10 };
-        
+
         // Using string length as key (all same length)
         ints.AsValueEnumerable().MaxBy(x => x.ToString()).ShouldBe(ints.MaxBy(x => x.ToString()));
         ints.ToValueEnumerable().MaxBy(x => x.ToString()).ShouldBe(ints.MaxBy(x => x.ToString()));
@@ -57,7 +57,7 @@ public class MaxByTest
     public void StringSourceWithIntegerKey()
     {
         var strings = new[] { "apple", "banana", "kiwi", "pineapple", "grape" };
-        
+
         // Using string length as key
         strings.AsValueEnumerable().MaxBy(x => x.Length).ShouldBe(strings.MaxBy(x => x.Length));
         strings.ToValueEnumerable().MaxBy(x => x.Length).ShouldBe(strings.MaxBy(x => x.Length));
@@ -67,11 +67,11 @@ public class MaxByTest
     public void StringSourceWithStringKey()
     {
         var strings = new[] { "apple", "banana", "kiwi", "pineapple", "grape" };
-        
+
         // Using the string itself as key
         strings.AsValueEnumerable().MaxBy(x => x).ShouldBe(strings.MaxBy(x => x));
         strings.ToValueEnumerable().MaxBy(x => x).ShouldBe(strings.MaxBy(x => x));
-        
+
         // Using first character as key
         strings.AsValueEnumerable().MaxBy(x => x.First()).ShouldBe(strings.MaxBy(x => x.First()));
         strings.ToValueEnumerable().MaxBy(x => x.First()).ShouldBe(strings.MaxBy(x => x.First()));
@@ -94,18 +94,18 @@ public class MaxByTest
     public void NullHandling()
     {
         var stringsWithNull = new[] { "banana", null, "apple", "cherry", null };
-        
+
         // Key selector that handles nulls
         stringsWithNull.AsValueEnumerable().MaxBy(x => x ?? "").ShouldBe(stringsWithNull.MaxBy(x => x ?? ""));
         stringsWithNull.ToValueEnumerable().MaxBy(x => x ?? "").ShouldBe(stringsWithNull.MaxBy(x => x ?? ""));
-        
+
         // Test when keys are null
-        var objects = new[] { 
+        var objects = new[] {
             new { Name = "A", Value = "High" },
             new { Name = "B", Value = null as string },
             new { Name = "C", Value = "Medium" }
         };
-        
+
         // Should return first non-null key's item or first item if all keys are null
         objects.AsValueEnumerable().MaxBy(x => x.Value).ShouldBe(objects.MaxBy(x => x.Value));
         objects.ToValueEnumerable().MaxBy(x => x.Value).ShouldBe(objects.MaxBy(x => x.Value));
@@ -128,16 +128,16 @@ public class MaxByTest
     public void TieBreaking()
     {
         // When multiple elements have the same max key, the first one should be returned
-        var items = new[] { 
+        var items = new[] {
             new { Name = "A", Value = 5 },
             new { Name = "B", Value = 10 },
             new { Name = "C", Value = 10 },  // Same value as B
             new { Name = "D", Value = 3 }
         };
-        
+
         var expected = items.MaxBy(x => x.Value);
         var actual = items.AsValueEnumerable().MaxBy(x => x.Value);
-        
+
         actual.ShouldBe(expected);
         actual.Name.ShouldBe("B"); // Should be "B" not "C"
     }
@@ -167,15 +167,15 @@ public class MaxByTest
     {
         var random = new Random(42);
         var data = new int[1000];
-        
+
         for (int i = 0; i < data.Length; i++)
         {
             data[i] = random.Next(-10000, 10000);
         }
-        
+
         var expected = data.MaxBy(x => x);
         var actual = data.AsValueEnumerable().MaxBy(x => x);
-        
+
         actual.ShouldBe(expected);
     }
 
@@ -183,7 +183,7 @@ public class MaxByTest
     {
         public string Name { get; }
         public int Age { get; }
-        
+
         public Person(string name, int age)
         {
             Name = name;