This PR add perf benchmarks that based on following `dotnet/performance` repository code.
https://github.com/dotnet/performance/blob/main/src/benchmarks/micro/runtime/Linq/Linq.cs

And comparing performance between System.Linq/ZLinq.

As far as I've confirmed benchmark results.
ZLinq seems run faster and memory efficient on almost cases. (Except for OrderBy benchmarks)

## Benchmark results on my environments.

### AggregateBy00
| Method                         | Mean    | Error    | StdDev   | Gen0         | Allocated |
|------------------------------- |--------:|---------:|---------:|-------------:|----------:|
| Linq_AggregateBy_Sum           | 2.630 s | 0.3650 s | 0.0200 s |  346000.0000 |   1.35 GB |
| Linq_GroupBy_ToDictionary_Sum  | 4.300 s | 0.3288 s | 0.0180 s | 1159000.0000 |   4.52 GB |
|                                |         |          |          |              |           |
| ZLinq_GroupBy_ToDictionary_Sum | 4.208 s | 0.1527 s | 0.0084 s |  713000.0000 |   2.78 GB |

Note: [AggregateBy->Sum method chain seems not supported by ZLinq](https://github.com/Cysharp/ZLinq/blob/39192be4181e314e07cac7cd0a993d70f98f7f06/src/ZLinq/Linq/Sum.cs#L7-L11).

####  Count00 benchmarks
| Method | Mean     | Error    | StdDev   | Allocated |
|------- |---------:|---------:|---------:|----------:|
| Linq   | 38.88 ms | 3.838 ms | 0.210 ms |      31 B |
| ZLinq  | 17.66 ms | 0.204 ms | 0.011 ms |      12 B |

### CountBy00 benchmarks
| Method                            | Mean    | Error    | StdDev   | Gen0         | Allocated  |
|---------------------------------- |--------:|---------:|---------:|-------------:|-----------:|
| Linq_CountBy_Count                | 1.725 s | 0.0592 s | 0.0032 s |  279000.0000 | 1113.89 MB |
| Linq_AggregateBy_Count            | 1.641 s | 0.1488 s | 0.0082 s |  285000.0000 | 1136.78 MB |
| Linq_GroupBy_ToDictionary_Count   | 3.049 s | 0.1411 s | 0.0077 s | 1031000.0000 | 4112.24 MB |
| Linq_ToLookup_ToDictionary_Count  | 2.904 s | 0.2721 s | 0.0149 s |  893000.0000 | 3562.93 MB |
|                                   |         |          |          |              |            |
| ZLinq_CountBy_Count               | 1.341 s | 0.0293 s | 0.0016 s |  237000.0000 |  946.05 MB |
| ZLinq_AggregateBy_Count           | 1.406 s | 0.0561 s | 0.0031 s |  237000.0000 |  946.05 MB |
| ZLinq_GroupBy_ToDictionary_Count  | 2.748 s | 0.0840 s | 0.0046 s |  560000.0000 | 2235.41 MB |
| ZLinq_ToLookup_ToDictionary_Count | 2.689 s | 0.5797 s | 0.0318 s |  614000.0000 | 2449.04 MB |

### GroupBy00 benchmarks
| Method              | Mean    | Error    | StdDev   | Gen0        | Allocated |
|-------------------- |--------:|---------:|---------:|------------:|----------:|
| Linq_GroupBy_Count  | 2.620 s | 0.4523 s | 0.0248 s | 793000.0000 |   3.09 GB |
| ZLinq_GroupBy_Count | 2.213 s | 0.0446 s | 0.0024 s | 332000.0000 |    1.3 GB |

### Order00 benchmarks
| Method                                  | Mean      | Error    | StdDev   | Gen0       | Allocated |
|---------------------------------------- |----------:|---------:|---------:|-----------:|----------:|
| Linq_OrderByDescending_Count_ElementAt  |  87.30 ms | 2.173 ms | 0.119 ms | 18285.7143 |  73.43 MB |
| ZLinq_OrderByDescending_Count_ElementAt | 129.20 ms | 4.548 ms | 0.249 ms |  4250.0000 |  17.55 MB |

### Where00 benchmarks
| Method | Mean      | Error     | StdDev   | Gen0       | Allocated  |
|------- |----------:|----------:|---------:|-----------:|-----------:|
| Linq   | 329.66 ms | 21.520 ms | 1.180 ms | 17000.0000 | 72000400 B |
| ZLinq  |  83.78 ms |  5.946 ms | 0.326 ms |          - |       67 B |

### Where01 benchmarks
| Method         | Mean     | Error     | StdDev   | Gen0       | Allocated  |
|--------------- |---------:|----------:|---------:|-----------:|-----------:|
| Linq_Combined  | 203.5 ms |  21.37 ms |  1.17 ms |  4000.0000 | 18000133 B |
| Linq_Nested    | 235.6 ms |   7.24 ms |  0.40 ms | 14333.3333 | 60000133 B |
|                |          |           |          |            |            |
| ZLinq_Combined | 140.0 ms | 619.87 ms | 33.98 ms |          - |       80 B |
| ZLinq_Nested   | 151.9 ms | 311.11 ms | 17.05 ms |          - |       80 B |