@neuecc 2025-03-18
Thank you.
Since our goal is to win in all benchmarks, your additions are very helpful.
Regarding AggregateSum, I plan to implement Sum(selector) now that the Source Generator constraints have been removed.