@neuecc 2025-03-18
I've checked the OrderBy results and it's due to the difference in ElementAt optimization.
System.Linq has optimization paths for First/Last/ElementAt, but in ZLinq, optimizations are only applied when a Span is available, and there are no dedicated paths for First/Last/ElementAt in other cases.
This is an intentional design decision (there are too few effective cases compared to how much it would increase code size).
To be honest, I don't think it's worth adding a dedicated path for ElementAt just to win in benchmarks.