@filzrev 2025-03-18
I've accidently committed test code that call `.ToArray()` after OrderByDescending(). 
(It should be removed before merge)

https://github.com/Cysharp/ZLinq/pull/29/files#diff-827a191ec392f456c85ed1fec3a785f91679cc10c5413597069970cea58a728fR41

If ToArray() are removed. Following benchmark results are obtained.

### Order00 benchmarks
| Method                                  | Mean      | Error    | StdDev   | Gen0       | Allocated |
|---------------------------------------- |----------:|---------:|---------:|-----------:|----------:|
| Linq_OrderByDescending_Count_ElementAt  |  47.35 ms | 2.426 ms | 0.133 ms | 14200.0000 |  56.65 MB |
| ZLinq_OrderByDescending_Count_ElementAt | 136.50 ms | 1.826 ms | 0.100 ms |  4250.0000 |  17.55 MB |