@filzrev 2025-03-18
I've removed `ToArray()` test code that are used for measuring `OrderByDescending` standalone performance.