@neuecc 2025-03-18
I'm wondering why GetProductList and similar operations are included within the benchmark rather than being done in Setup.
It would also add noise to the measurements, so while it wouldn't be an exact copy from dotnet/performance anymore, I might make changes to move these operations to the Setup method.