@filzrev 2025-03-18
> I'm wondering why GetProductList and similar operations are included within the benchmark rather than being done in Setup.

I'll create another PR to fixed above issue.

Current benchmark is executed with `ShortRunJob` and default `Throughput` RunStrategy.
So it's not expected to be affect current benchmark results.
Because warmup call is executed invoked 3 times. 

It'll be required when running benchmark with `RunStrategy.ColdStart`.